--- pylot/plugins/a__init__.py
@@ -1,1 +1,0 @@
-from .helpers.pylot_helpers import PyLOTHelpers
