--- pylot/plugins/config.json
@@ -1,186 +1,0 @@
-{
-  "plugins": [
-    "pylot.plugins.get_cumulus_records",
-    "pylot.plugins.get_status",
-    "pylot.plugins.manage_cumulus_records"
-  ],
-  "options": [
-    {
-      "prog": {
-        "name": "get_status",
-        "flags": [
-          {
-            "name_or_flags": "-f, --filter",
-            "dest": "filter",
-            "nargs": "+",
-            "help": "filter the returned data (where x = y)"
-          },
-          {
-            "name_or_flags": "-s, --select",
-            "dest": "fields",
-            "nargs": "+",
-            "help": "select some fields from the requested data"
-          },
-          {
-            "name_or_flags": "-o, --output",
-            "dest": "output_format",
-            "nargs": 1,
-            "help": "chose output format",
-            "choices": [
-              "json",
-              "table"
-            ],
-            "default": [
-              "table"
-            ]
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "get_cumulus_records",
-        "flags": [
-          {
-            "name_or_flags": "-t, --type",
-            "dest": "record_type",
-            "required": true,
-            "choices": [
-              "collections",
-              "granules",
-              "providers",
-              "rules"
-            ],
-            "nargs": 1,
-            "help": "Record type (Granule| Granules|Collections)"
-          },
-          {
-            "name_or_flags": "-f, --filter",
-            "dest": "filter",
-            "nargs": "+",
-            "help": "Filter the returned data (where x = y)"
-          },
-          {
-            "name_or_flags": "-s, --select",
-            "dest": "fields",
-            "nargs": "+",
-            "help": "Select some fields from the requested data"
-          },
-          {
-            "name_or_flags": "-l, --limit",
-            "dest": "limit",
-            "nargs": 1,
-            "default": [
-              10
-            ],
-            "help": "Limit the requested data (default to 10). If 0 is provided with an OpenSearch query an attempt will be made to return all available results. This could take a while."
-          },
-          {
-            "name_or_flags": "-c, --count",
-            "dest": "count",
-            "nargs": 1,
-            "default": [
-              "false"
-            ],
-            "choices": [
-              "true",
-              "false"
-            ],
-            "help": "Get ONLY the count of the result"
-          },
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "query_data",
-            "required": false,
-            "nargs": 1,
-            "help": "JSON file containing an OpenSearch query"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "update_provider",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "provider_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing provider definition"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "update_collection",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "collection_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing collection definition"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "update_granule",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "granule_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing granule definition"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "create_provider",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "provider_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing provider definition"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "create_collection",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "collection_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing collection definition"
-          }
-        ]
-      }
-    },
-    {
-      "prog": {
-        "name": "create_granule",
-        "flags": [
-          {
-            "name_or_flags": "-d, --data",
-            "dest": "granule_data",
-            "required": true,
-            "nargs": 1,
-            "help": "JSON file containing granule definition"
-          }
-        ]
-      }
-    }
-  ]
-}
