+++ pylot/test_plugins/cumulus_api/tests/fixtures/certificate.pfx
